move: showcase -> piece
