+++ frontend/src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Shop from "./pages/shop/Shop";
 import Home from "./pages/home/Home";
 import Cart from "./pages/cart/Cart";
-import Showcase from "./pages/showcase/Showcase";
+import Showcase from "./pages/showcase/Piece";
 import Checkout from "./pages/checkout/Checkout";
 import ErrorPage from "./pages/ErrorPage";
 import { PieceService } from "./services/PieceService";
