+++ frontend/src/pages/showcase/Piece.tsx
@@ -3,21 +3,21 @@ import Footer from "../../components/Footer";
 import Header from "../../components/Header";
 import SecondaryHeroSection from "../../components/SecondaryHeroSection";
 import SeperationSection from "../../components/SeperationSection";
-import ShowcaseSection from "../../components/ShowcaseSection";
+import ShowcaseSection from "../../components/showcase/ShowcaseSection";
 import { PieceDTO } from "../../dtos/dtos";
 
 const Shop = ({ piece }: { piece: PieceDTO }) => {
   return (
     <body className="flex flex-col  bg-black">
       <Header />
-      <SecondaryHeroSection
-        backdrop={true}
-        image={
-          "../../../public/albums/" +
-          piece.title.replace(/ /g, "-").toLowerCase() +
-          ".png"
-        }
-      />
+      {/* <SecondaryHeroSection */}
+      {/*   backdrop={true} */}
+      {/*   image={ */}
+      {/*     "../../../public/albums/" + */}
+      {/*     piece.title.replace(/ /g, "-").toLowerCase() + */}
+      {/*     ".png" */}
+      {/*   } */}
+      {/* /> */}
       <ShowcaseSection piece={piece} />
       <SeperationSection />
       <ContactSection />
